Move updateOrder event template to directory
