+++ src/constants/MoveHistory/EventTemplates/index.js
@@ -36,7 +36,7 @@ export { default as updateMTOShipmentAgent } from './updateMTOShipmentAgent';
 export { default as updateMTOShipmentDeprecatePaymentRequest } from './updateMTOShipmentDeprecatePaymentRequest';
 export { default as reweighPaymentRequest } from './reweighPaymentRequest';
 export { default as updateMTOServiceItemAddress } from './CreateMTOServiceItem/updateMTOServiceItemAddress';
-export { default as updateOrder } from './updateOrder';
+export { default as updateOrder } from './updateOrders/updateOrder';
 export { default as updatePaymentRequest } from './updatePaymentRequest';
 export { default as updatePaymentRequestJobRunner } from './updatePaymentRequestJobRunner';
 export { default as updatePaymentRequestStatus } from './updatePaymentRequestStatus';
